machines/5i,vi: fix ARM/MIPS register fields being long instead of a real 32-bit type
Both emulators' Registers.r[]/cc1/cc2 were long (64-bit on this host)
even though ARM and MIPS registers are 32 bits. That let several ALU
paths compute a 32-bit result using a shift or comparison that only
stays correct if the type itself is 32 bits:

- 5i: Idp3()'s rotated-immediate operand ("o2 << (32-sc)") overflowed
  past bit 31 for small rotate amounts instead of wrapping, corrupting
  the destination register -- concretely, a plain `SUB $8,R13,R13`
  (any stack-adjusting call prologue) silently produced a wild SP,
  crashing 5i with a TLB miss on the next store. shift()'s ASR/rotate
  branches had the same latent issue, now fixed for free by the type
  change (plus an explicit sign-extend for ASR, since v is unsigned
  now). runcmp()'s mi/pl/ge/lt/gt/le condition codes needed matching
  updates since cc1/cc2 are unsigned now too.
- vi: Ilw() (load word) assigned getmem_w()'s u32int return into the
  long register field, zero- instead of sign-extending a negative
  32-bit value -- so Sslt()'s plain signed `<` comparison saw e.g.
  0xfffffffe as a huge positive number instead of -2, turning a
  bounded loop into an infinite one.

Both were reachable through the most ordinary code: any print() call
with a format specifier hit the ASR/rotate path on ARM and an SLT-based
loop bound check on MIPS, so lib_core/libc's GOOS=plan9 support (added
in e8421dcf0) could run a plain print() but not print("...%d...",n) --
blocking tests/c/hello_libc's hello.c/io.c from ever running under
5i/vi. Verified via real 5i/vi execution: tests/5i/smoke.sh,
tests/vi/smoke.sh, and hello.c now pass end to end on both arches
(io.c's open/read/seek/write path surfaced separate, pre-existing
syscall-emulation gaps -- out of scope here, left for a follow-up).

+++ machines/vi/mips.h
@@ -92,7 +92,21 @@ struct Registers
 	// getmem_w() in mem.c for where the sign-extension itself happens)
 	u32int	ir;
 	Inst	*ip;
-	long	r[32];
+	// claude: int32, not long -- same "code assumed a 32-bit host"
+	// bug as reg.ir above, but hitting signed comparisons instead of
+	// instruction dispatch: a MIPS register is 32 bits, and loads
+	// like Ilw() assign a u32int (getmem_w()'s return type) into this
+	// field, which zero-extends into a 64-bit `long` instead of
+	// sign-extending -- so a loaded word like 0xfffffffe (meant to be
+	// -2) read back as a huge *positive* 64-bit value, making Sslt()'s
+	// plain `<` comparison (and any other signed op relying on this
+	// field's own sign) wrong. Ssltu()'s existing explicit
+	// `(unsigned)reg.r[rs]` cast for the *unsigned* comparison is the
+	// tell that this field was always meant to be exactly 32 bits
+	// wide, signed by default, with unsigned ops opting in via a cast
+	// -- not a 64-bit field holding an unspecified mix of sign- and
+	// zero-extended values.
+	int32	r[32];
 	ulong	mhi;
 	ulong	mlo;
